+++ webapp/src/views/projects/dashboard/ProjectTotals.tsx
@@ -198,6 +198,7 @@ export const ProjectTotals: React.FC<{
   const tagsClickable = tagsPresent && canViewKeys;
 
   const membersAccessible = config.authentication && canViewMembers;
+  const membersEditable = membersAccessible && canEditMembers;
 
   return (
     <>
@@ -302,7 +303,7 @@ export const ProjectTotals: React.FC<{
               {t('project_dashboard_member_count', 'Members')}
             </StyledTileDescription>
           </StyledTileDataItem>
-          {canEditMembers && (
+          {membersEditable && (
             <StyledTileEdit>
               <Edit fontSize="small" />
             </StyledTileEdit>